+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 3. Added color parameters to `Tooltip`
 4. Created `StringAlignUtils`
 5. Added option to not render gray background in ScreenUIRenderer
+6. Added `sizeBehaviour` field to `Container`
 
 # v1.0.5 - Minor Update
 ## Built-in Components